refactor(tests): split into unit and integration tests folder
