+++ tests/intergration/test_client_and_ops.py
