+++ tests/unit/test_host_applications.py
