+++ tests/unit/test_registering_base.py
