+++ tests/unit/test_structural.py
