+++ tests/unit/test_traverse_value.py
@@ -4,19 +4,19 @@ from specklepy.objects.base import Base
 from specklepy.serialization.base_object_serializer import BaseObjectSerializer
 
 
-class TestBase(Base):
+class FakeBase(Base):
     foo: List[str]
     bar: int
 
 
 def test_traverse_value():
-    base = TestBase(bar=1)
+    base = FakeBase(bar=1)
     base.foo = [None]
     serializer = BaseObjectSerializer()
     object_id, object_dict = serializer.traverse_base(base)
     assert object_dict == {
         "id": object_id,
-        "speckle_type": "TestBase",
+        "speckle_type": "FakeBase",
         "applicationId": None,
         "foo": [None],
         "units": None,
